wrap use wallet with context
